todo unit test

+++ todo/src/routes/delete.ts
@@ -1,15 +1,21 @@
 import express, { Response, Request } from 'express'
-import { validateRequest, requireAuth, NotFoundError, NotAuthorizedError } from '@fan-todo/common'
+import { requireAuth, NotFoundError, NotAuthorizedError } from '@fan-todo/common'
 
 import { Todo } from '../models/todo'
 
 const router = express.Router()
 
-router.delete('/api/todo/:id', async (req: Request, res: Response) => {
+router.delete('/api/todo/:id', requireAuth, async (req: Request, res: Response) => {
   const { id } = req.params
-  await Todo.findByIdAndDelete(id)
-
-  res.status(200).send({})
+  const todo = await Todo.findById(id)
+  if (!todo) {
+    throw new NotFoundError()
+  }
+  if (todo.userId !== req.currentUser!.id) {
+    throw new NotAuthorizedError()
+  }
+  const response = await Todo.findByIdAndDelete(id)
+  res.status(200).send(response)
 })
 
 export { router as deleteTodoRouter }
